+++ packages/sdk/src/crud/keys.ts
@@ -1,0 +1,16 @@
+import { MCPClient } from "../fetcher.ts";
+
+/**
+ * Create an API key
+ * @param workspace - The workspace
+ * @param params - API key parameters
+ * @returns The created API key
+ */
+export const createAPIKey = (
+  workspace: string,
+  params: {
+    claims?: Record<string, unknown>;
+    name: string;
+    policies: Array<{ effect: "allow" | "deny"; resource: string }>;
+  },
+) => MCPClient.forWorkspace(workspace).API_KEYS_CREATE(params);
